+++ src/utils/misc.cpp
@@ -1076,3 +1076,26 @@ void Utils::Misc::printInfo(QString text) {
     printf("%s\n", text.toLatin1().data());
 #endif
 }
+
+/**
+ * Converts a byte size to a human readable size
+ *
+ * @param size
+ * @return
+ */
+QString Utils::Misc::toHumanReadableByteSize(qint64 size)
+{
+    float num = size;
+    QStringList list;
+    list << "KB" << "MB" << "GB" << "TB";
+
+    QStringListIterator i(list);
+    QString unit("bytes");
+
+    while (num >= 1024.0 && i.hasNext()) {
+        unit = i.next();
+        num /= 1024.0;
+    }
+
+    return QString().setNum(num, 'f', 2) + " " + unit;
+}